refactor: . Move src/Query/Sort* to src/Query/Sort/
