+++ src/Query/Filter/DateField.ts
@@ -3,7 +3,7 @@ import { DateRange } from '../DateRange';
 import type { Task } from '../../Task';
 import { DateParser } from '../DateParser';
 import { Explanation } from '../Explain/Explanation';
-import type { Comparator } from '../Sorter';
+import type { Comparator } from '../Sort/Sorter';
 import { compareByDate } from '../../lib/DateTools';
 import type { GrouperFunction } from '../Group/Grouper';
 import { TemplatingPluginTools } from '../../lib/TemplatingPluginTools';
