+++ src/Query/Filter/DescriptionField.ts
@@ -1,6 +1,6 @@
 import { GlobalFilter } from '../../Config/GlobalFilter';
 import type { Task } from '../../Task';
-import type { Comparator } from '../Sorter';
+import type { Comparator } from '../Sort/Sorter';
 import { TextField } from './TextField';
 
 /**
