+++ src/Query/Filter/Field.ts
@@ -1,5 +1,5 @@
-import { Sorter } from '../Sorter';
-import type { Comparator } from '../Sorter';
+import { Sorter } from '../Sort/Sorter';
+import type { Comparator } from '../Sort/Sorter';
 import * as RegExpTools from '../../lib/RegExpTools';
 import { Grouper } from '../Group/Grouper';
 import type { GrouperFunction } from '../Group/Grouper';
