+++ src/Query/Filter/PriorityField.ts
@@ -1,6 +1,6 @@
 import { Priority, Task } from '../../Task';
 import { Explanation } from '../Explain/Explanation';
-import type { Comparator } from '../Sorter';
+import type { Comparator } from '../Sort/Sorter';
 import type { GrouperFunction } from '../Group/Grouper';
 import { Field } from './Field';
 import { Filter } from './Filter';
