+++ src/Query/Filter/RecurringField.ts
@@ -1,6 +1,6 @@
 import type { Task } from '../../Task';
 import type { GrouperFunction } from '../Group/Grouper';
-import type { Comparator } from '../Sorter';
+import type { Comparator } from '../Sort/Sorter';
 import { FilterInstructionsBasedField } from './FilterInstructionsBasedField';
 
 export class RecurringField extends FilterInstructionsBasedField {
