+++ src/Query/Filter/StatusField.ts
@@ -1,5 +1,5 @@
 import type { Task } from '../../Task';
-import type { Comparator } from '../Sorter';
+import type { Comparator } from '../Sort/Sorter';
 import type { GrouperFunction } from '../Group/Grouper';
 import { FilterInstructionsBasedField } from './FilterInstructionsBasedField';
