+++ src/Query/Filter/TagsField.ts
@@ -1,6 +1,6 @@
 import type { Task } from '../../Task';
-import type { Comparator } from '../Sorter';
-import { Sorter } from '../Sorter';
+import type { Comparator } from '../Sort/Sorter';
+import { Sorter } from '../Sort/Sorter';
 import type { GrouperFunction } from '../Group/Grouper';
 import type { FilterOrErrorMessage } from './FilterOrErrorMessage';
 import { FilterInstructions } from './FilterInstructions';
