+++ src/Query/Filter/TextField.ts
@@ -2,7 +2,7 @@ import type { Task } from '../../Task';
 import { SubstringMatcher } from '../Matchers/SubstringMatcher';
 import { RegexMatcher } from '../Matchers/RegexMatcher';
 import type { IStringMatcher } from '../Matchers/IStringMatcher';
-import type { Comparator } from '../Sorter';
+import type { Comparator } from '../Sort/Sorter';
 import type { GrouperFunction } from '../Group/Grouper';
 import { errorMessageForException } from '../../lib/ExceptionTools';
 import { Field } from './Field';
