+++ src/Query/FilterParser.ts
@@ -22,7 +22,7 @@ import { StatusTypeField } from './Filter/StatusTypeField';
 
 import { RecurrenceField } from './Filter/RecurrenceField';
 import type { FilterOrErrorMessage } from './Filter/FilterOrErrorMessage';
-import type { Sorter } from './Sorter';
+import type { Sorter } from './Sort/Sorter';
 import type { Grouper } from './Group/Grouper';
 import { FolderField } from './Filter/FolderField';
 import { RootField } from './Filter/RootField';
