+++ src/Query/Query.ts
@@ -7,8 +7,8 @@ import type { IQuery } from '../IQuery';
 import { getSettings } from '../Config/Settings';
 import { errorMessageForException } from '../lib/ExceptionTools';
 import { logging } from '../lib/logging';
-import { Sort } from './Sort';
-import type { Sorter } from './Sorter';
+import { Sort } from './Sort/Sort';
+import type { Sorter } from './Sort/Sorter';
 import { TaskGroups } from './Group/TaskGroups';
 import * as FilterParser from './FilterParser';
 import type { Grouper } from './Group/Grouper';
