+++ tests/TestingTools/SortingTestHelpers.ts
@@ -1,6 +1,6 @@
-import type { Sorter } from '../../src/Query/Sorter';
+import type { Sorter } from '../../src/Query/Sort/Sorter';
 import type { Task } from '../../src/Task';
-import { Sort } from '../../src/Query/Sort';
+import { Sort } from '../../src/Query/Sort/Sort';
 import { SearchInfo } from '../../src/Query/SearchInfo';
 
 export function sortBy(sorters: Sorter[], tasks: Task[]) {
